Refactor cities-of-brazil actions
