+++ .env.example
@@ -1,2 +1,1 @@
-OGH_GITEA_HOST_URL=
 GITEA_ACCESS_TOKEN=
